requirements.txt
streamlit==1.31.1
pandas==2.2.0
beautifulsoup4==4.12.3
openai>=1.70.0
requests==2.31.0
python-dotenv==1.0.0 